Moving secureTransfer to spring.cloud.azure.storage
